no array

+++ controllers/limitController.js
@@ -1,19 +1,24 @@
-// limitController.js
-
 const axios = require('axios');
+const Group = require('../models/groupModel'); // Assicurati che il percorso sia corretto
 
 // Controller per creare un limite generico per il gruppo
 exports.createLimitGeneric = async (req, res, next) => {
   try {
     const { chatId, limit } = req.body;
-    console.log('chatId', chatId);
-    console.log('limit', limit);
     if (!chatId || !limit) {
       return res.status(400).json({ error: 'chatId e limit sono richiesti.' });
     }
-    console.log('process.env.BOT_API_URL', process.env.BOT_API_URL);
+
+    // Aggiorna il limite generico nel bot
     const endpoint = `${process.env.BOT_API_URL}/groupLimitGeneric`;
     const response = await axios.post(endpoint, { chatId, limit });
+
+    // Aggiorna il modello Group con il nuovo limite
+    await Group.findOneAndUpdate(
+      { groupId: chatId },
+      { $addToSet: { limits: limit } }, // Aggiungi il limite se non è già presente
+      { new: true, upsert: true }, // Crea un nuovo documento se non esiste
+    );
 
     res.status(response.status).json(response.data);
   } catch (error) {
@@ -22,18 +27,28 @@ exports.createLimitGeneric = async (req, res, next) => {
   }
 };
 
-// Controller per cancellare un limite generico per il gruppo
+// Controller per cancellare tutti i limiti generici per il gruppo
 exports.deleteLimitGeneric = async (req, res, next) => {
   try {
     const { chatId } = req.params;
-    console.log('************************************chatId', chatId);
+    if (!chatId) {
+      return res.status(400).json({ error: 'chatId è richiesto.' });
+    }
+
+    // Rimuovi i limiti generici dal bot
     const endpoint = `${process.env.BOT_API_URL}/groupLimitGeneric/${chatId}`;
     const response = await axios.delete(endpoint);
+
+    // Rimuovi tutti i limiti dal modello Group
+    await Group.findOneAndUpdate(
+      { groupId: chatId },
+      { $set: { limits: [] } }, // Rimuovi tutti i limiti
+    );
 
     res.status(response.status).json(response.data);
   } catch (error) {
     console.error(
-      'Errore durante la cancellazione del limite generico:',
+      'Errore durante la cancellazione dei limiti generici:',
       error,
     );
     next(error);
